turn chunks into arrays only once

+++ tests/test_chunked.py
@@ -56,7 +56,7 @@ class TestChunked(unittest.TestCase):
 
     def test_chunked_dtype(self):
         a = ChunkedArray([[], [0, 1, 2, 3, 4], [5, 6], [], [7, 8, 9], []])
-        self.assertEqual(a.dtype, numpy.dtype(float))
+        self.assertEqual(a.dtype, numpy.dtype(int))
 
         a = ChunkedArray([[0, 1, 2, 3, 4], [5, 6], [], [7, 8, 9], []])
         self.assertEqual(a.dtype, numpy.dtype(int))
